Update patch set 1
Patch Set 1: Code-Review-1

Ya,  this may not even be needed.  According to some reports, IR blaster was working as-is for people testing CM-11.0.   I have not personally tested it with or without anything specific, as I think this is sort of on the back-burner for now,  camera and sensors are more important.

Patch-set: 1
Reviewer: Gerrit User 1000124 <1000124@56841c99-7c94-43d4-bf0b-cb4707e66d55>
Label: Code-Review=-1