Update patch set 1
Patch Set 1:

So here is the alternative for you - you can implement stuff that tweaks kernel stuff like a profile also in device settings. You dont need a power hint for that plus it keeps AOSP code clean. Any UI element for that would be in device settings anyway.

Patch-set: 1